+++ hera/debugger/minilanguage.py
@@ -18,7 +18,7 @@ class MiniParser:
 
       start := expr
 
-      expr := mem | REGISTER | INT
+      expr := mem | REGISTER | INT | MINUS expr
       mem  := MEM expr RBRACKET
     """
 
@@ -44,6 +44,8 @@ class MiniParser:
                 return IntNode(int(tkn[1], base=0))
             except ValueError:
                 raise SyntaxError("invalid integer literal: {}".format(tkn[1]))
+        elif tkn[0] == TOKEN_MINUS:
+            return MinusNode(self.match_expr())
         elif tkn[0] == TOKEN_REGISTER:
             return RegisterNode(tkn[1])
         elif tkn[0] == TOKEN_SYMBOL:
@@ -74,6 +76,7 @@ RegisterNode = namedtuple("RegisterNode", ["value"])
 IntNode = namedtuple("IntNode", ["value"])
 BoolNode = namedtuple("BoolNode", ["value"])
 SymbolNode = namedtuple("SymbolNode", ["value"])
+MinusNode = namedtuple("MinusNode", ["arg"])
 
 
 class MiniLexer:
@@ -107,10 +110,7 @@ class MiniLexer:
             length = self.read_int()
             return self.advance_and_return(TOKEN_INT, length=length)
         elif ch == "-":
-            self.position += 1
-            length = self.read_int()
-            self.position -= 1
-            return self.advance_and_return(TOKEN_INT, length=length)
+            return self.advance_and_return(TOKEN_MINUS)
         elif ch == "#" and self.peek() in "tfTF":
             if self.peek() in "tT":
                 return self.advance_and_return(TOKEN_TRUE, length=2)
@@ -182,11 +182,11 @@ class MiniLexer:
 TOKEN_INT = "TOKEN_INT"
 TOKEN_MEM = "TOKEN_MEM"
 TOKEN_REGISTER = "TOKEN_REGISTER"
-TOKEN_LBRACKET = "TOKEN_LBRACKET"
 TOKEN_RBRACKET = "TOKEN_RBRACKET"
 TOKEN_SYMBOL = "TOKEN_SYMBOL"
 TOKEN_TRUE = "TOKEN_TRUE"
 TOKEN_FALSE = "TOKEN_FALSE"
+TOKEN_MINUS = "TOKEN_MINUS"
 TOKEN_EOF = "TOKEN_EOF"
 TOKEN_UNKNOWN = "TOKEN_UNKNOWN"