+++ test/test_unit/test_minilanguage.py
@@ -4,16 +4,19 @@ from hera.debugger.minilanguage import (
     MemoryNode,
     MiniLexer,
     MiniParser,
+    MinusNode,
     RegisterNode,
     SymbolNode,
     TOKEN_EOF,
     TOKEN_FALSE,
     TOKEN_INT,
     TOKEN_MEM,
+    TOKEN_MINUS,
     TOKEN_RBRACKET,
     TOKEN_REGISTER,
     TOKEN_SYMBOL,
     TOKEN_TRUE,
+    TOKEN_UNKNOWN,
 )
 
 
@@ -27,7 +30,7 @@ def test_lex_mini_language_with_small_example():
 
 def test_lex_mini_language_with_big_example():
     # This isn't a syntactically valid expression, but it doesn't matter to the lexer.
-    lexer = MiniLexer("M[FP_alt] R15 0xabc some_symbol #t #F")
+    lexer = MiniLexer("M[FP_alt] R15 0xabc some_symbol #t #F -10 ?")
 
     assert lexer.next_token() == (TOKEN_MEM, "M[")
     assert lexer.next_token() == (TOKEN_REGISTER, "FP_alt")
@@ -37,6 +40,9 @@ def test_lex_mini_language_with_big_example():
     assert lexer.next_token() == (TOKEN_SYMBOL, "some_symbol")
     assert lexer.next_token() == (TOKEN_TRUE, "#t")
     assert lexer.next_token() == (TOKEN_FALSE, "#F")
+    assert lexer.next_token() == (TOKEN_MINUS, "-")
+    assert lexer.next_token() == (TOKEN_INT, "10")
+    assert lexer.next_token() == (TOKEN_UNKNOWN, "?")
     assert lexer.next_token() == (TOKEN_EOF, "")
     assert lexer.next_token() == (TOKEN_EOF, "")
 
@@ -62,13 +68,14 @@ def test_parse_memory_expression():
 
 
 def test_parse_memory_expression_with_integer():
-    parser = MiniParser(MiniLexer("M[0o12]"))
+    parser = MiniParser(MiniLexer("M[-0o12]"))
 
     tree = parser.parse()
 
     assert isinstance(tree, MemoryNode)
-    assert isinstance(tree.address, IntNode)
-    assert tree.address.value == 0o12
+    assert isinstance(tree.address, MinusNode)
+    assert isinstance(tree.address.arg, IntNode)
+    assert tree.address.arg.value == 0o12
 
 
 def test_parse_boolean():